+++ util/package.go
@@ -345,8 +345,11 @@ func (p *Package) LoadDataSets(packagePath string) error {
 			for iK, _ := range datasource.Inputs {
 				for _, stream := range d.Streams {
 					if stream.Input == p.Datasources[dK].Inputs[iK].Type {
+						if stream.File == "" {
+							stream.File = "stream.yml"
+						}
 						stream.Dataset = d.ID
-						streamTemplate := filepath.Join(datasetBasePath, "agent", "stream", "stream.yml")
+						streamTemplate := filepath.Join(datasetBasePath, "agent", "stream", stream.File)
 
 						streamTemplateData, err := ioutil.ReadFile(streamTemplate)
 						if err != nil {